requirements.txt
asgiref==3.7.2
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
cssselect2==0.7.0
Django==5.0.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
fonttools==4.47.2
h11==0.14.0
html5lib==1.1
idna==3.6
outcome==1.3.0.post0
pillow==10.2.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
PySocks==1.7.1
pytz==2023.4
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
weasyprint==60.2
webencodings==0.5.1
wsproto==1.2.0
zopfli==0.2.3
